Make NOAA binary grid reader more flexible
In the previous design, the reader would parse all dates present in a
file, optionally skipping over files that had already been parsed. As
the workflow has switched to a design in which each timestep is
responsible for acquiring and preparing its own inputs, it makes more
sense for the grid reader to accept a timestep as a command line
argument, and extract only the requested timestep.