Add ability to replace targets or dependencies with tags
This is useful in certain situations to eliminate rules with very large
number of implicit targets, such as a command that runs the LSM from
1948-present and generates a result and state file for each month.
(Rules with large numbers of implicit targets seem to cause at least
some versions of Make to hang when executing a workflow in parallel.) The
actual outputs from a given command can be replaced with a tag. A rule is
then written for each of the actual outputs, showing it to depend on the
tag.

This is somewhat dangerous, because nothing guarantees that the state of
the tag and the actual outputs stay in sync, but there doesn't seem to
be a better alternative for executing a WSIM workflow with Make.